application/requirements-backend.txt
fastapi
uvicorn
pytest
requests
wonderwords
aiofiles
Pillow
celery
jinja2
boto3
moto